# lab1/docs/swagger.yaml
basePath: /
definitions:
...
  ds.CompletePvlcMedCardRequest:
    properties:
      action:
        description: '"complete" или "reject"'
        type: string
    required:
    - action
    type: object
...
  ds.MedMmPvlcCalculationResponse:
    properties:
      description:
        type: string
      final_result:
        type: number
      formula:
        type: string
      image_url:
        type: string
      input_height:
        type: number
      pvlc_med_formula_id:
        type: integer
      title:
        type: string
    type: object
...
  ds.PvlcMedCardResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      doctor_name:
        type: string
      finalized_at:
        type: string
      id:
        type: integer
      med_calculations:
        items:
          $ref: '#/definitions/ds.MedMmPvlcCalculationResponse'
        type: array
      patient_name:
        type: string
      status:
        type: string
      total_result:
        type: number
    type: object
...
  ds.UpdatePvlcMedCardRequest:
    properties:
      doctor_name:
        type: string
      patient_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@lungcalc.ru
    name: API Support
    url: http://localhost:8080
  description: |-
    API для расчета должной жизненной емкости легких (ДЖЕЛ)
    Лабораторная работа 4 - Добавление аутентификации и авторизации
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Lung Capacity Calculation API
  version: "1.0"
paths:
...
  /api/pvlc-med-cards:
    get:
      description: Возвращает список заявок пользователя (для модераторов - все заявки)
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Фильтр по дате от
        in: query
        name: date_from
        type: string
      - description: Фильтр по дате до
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.PvlcMedCardResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение списка заявок
      tags:
      - medical-cards
  /api/pvlc-med-cards/{id}:
    delete:
      description: Удаляет заявку (только черновики и только владельцем)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление заявки
      tags:
      - medical-cards
    get:
      description: Возвращает информацию о конкретной заявке
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.PvlcMedCardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение конкретной заявки
      tags:
      - medical-cards
    put:
      consumes:
      - application/json
      description: Обновляет поля заявки (только для владельца)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.UpdatePvlcMedCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление заявки
      tags:
      - medical-cards
  /api/pvlc-med-cards/{id}/complete:
    put:
      consumes:
      - application/json
      description: Завершает или отклоняет заявку (только для модераторов)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: 'Действие: complete или reject'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.CompletePvlcMedCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Завершение/отклонение заявки
      tags:
      - medical-cards
  /api/pvlc-med-cards/{id}/form:
    put:
      description: Переводит заявку из статуса черновика в сформированную
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Формирование заявки
      tags:
      - medical-cards
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен для аутентификации. Формат: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"